+++ HISTORY.md
@@ -18,6 +18,11 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 -->
 
+## 0.2.252 - 2025-09-04
+
+- Rebuild detached tabs by cloning widgets instead of reparenting to avoid
+  cross-toplevel ``TclError``
+
 ## 0.2.251 - 2025-09-04
 
 - Reparent detached tabs across toplevel windows using native OS APIs so